update gcampustools models
