+++ gcampus/tools/models/__init__.py
@@ -13,4 +13,5 @@
 #  You should have received a copy of the GNU Affero General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from gcampus.tools.models.calibration import MeasurementKit, Calibration
+from gcampus.tools.models.kit import MeasurementKit
+from gcampus.tools.models.calibration import Calibration
